- improve type cleanness on gcc 4.8.4 (Ubuntu 4.8.4-2ubuntu1~14.04.3)

+++ nsd/tclobj.c
@@ -201,7 +201,7 @@ Ns_TclSetStringRep(Tcl_Obj *objPtr, const char *bytes, int length)
         length = (int)strlen(bytes);
     }
     objPtr->length = length;
-    objPtr->bytes = ckalloc((size_t) length + 1u);
+    objPtr->bytes = ckalloc((unsigned) length + 1u);
     memcpy(objPtr->bytes, bytes, (size_t) length + 1u);
 }
